backend/requirements.txt
asgiref==3.8.1
certifi==2020.6.20
chardet==3.0.4
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==23.0.0
idna==2.10
packaging==24.1
psycopg2==2.9.9
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==1.26.5
whitenoise==6.7.0
